+++ border-core/src/record.rs
@@ -8,11 +8,11 @@
 //!
 //! // following values are obtained with some process in reality
 //! let step = 1;
-//! let obs = vec![1f32, 2.0, 3.0, 4.0, 5.0];
-//! let reward = -1f32;
+//! let obs = vec![1f64, 2.0, 3.0, 4.0, 5.0];
+//! let reward = -1f64;
 //!
 //! let mut record = Record::empty();
-//! record.insert("Step", RecordValue::Scalar(step as f32));
+//! record.insert("Step", RecordValue::Scalar(step as f64));
 //! record.insert("Reward", RecordValue::Scalar(reward));
 //! record.insert("Obs", RecordValue::Array1(obs));
 //! ```
@@ -39,19 +39,19 @@ use crate::error::LrrError;
 /// Represents possible types of values in a [`Record`].
 pub enum RecordValue {
     /// Represents a scalar, e.g., optimization steps and loss value.
-    Scalar(f32),
+    Scalar(f64),
 
     /// Represents a datetime.
     DateTime(DateTime<Local>),
 
     /// A 1-dimensional array
-    Array1(Vec<f32>),
+    Array1(Vec<f64>),
 
     /// A 2-dimensional array
-    Array2(Vec<f32>, [usize; 2]),
+    Array2(Vec<f64>, [usize; 2]),
 
     /// A 3-dimensional array
-    Array3(Vec<f32>, [usize; 3]),
+    Array3(Vec<f64>, [usize; 3]),
 
     /// String
     String(String),
@@ -114,7 +114,7 @@ impl Record {
     /// Get scalar value.
     ///
     /// * `key` - The key of an entry in the record.
-    pub fn get_scalar(&self, k: &str) -> Result<f32, LrrError> {
+    pub fn get_scalar(&self, k: &str) -> Result<f64, LrrError> {
         if let Some(v) = self.0.get(k) {
             match v {
                 RecordValue::Scalar(v) => Ok(*v as _),
@@ -126,7 +126,7 @@ impl Record {
     }
 
     /// Get Array1 value.
-    pub fn get_array1(&self, k: &str) -> Result<Vec<f32>, LrrError> {
+    pub fn get_array1(&self, k: &str) -> Result<Vec<f64>, LrrError> {
         if let Some(v) = self.0.get(k) {
             match v {
                 RecordValue::Array1(v) => Ok(v.clone()),
@@ -138,7 +138,7 @@ impl Record {
     }
 
     /// Get Array2 value.
-    pub fn get_array2(&self, k: &str) -> Result<(Vec<f32>, [usize; 2]), LrrError> {
+    pub fn get_array2(&self, k: &str) -> Result<(Vec<f64>, [usize; 2]), LrrError> {
         if let Some(v) = self.0.get(k) {
             match v {
                 RecordValue::Array2(v, s) => Ok((v.clone(), s.clone())),
@@ -150,7 +150,7 @@ impl Record {
     }
 
     /// Get Array3 value.
-    pub fn get_array3(&self, k: &str) -> Result<(Vec<f32>, [usize; 3]), LrrError> {
+    pub fn get_array3(&self, k: &str) -> Result<(Vec<f64>, [usize; 3]), LrrError> {
         if let Some(v) = self.0.get(k) {
             match v {
                 RecordValue::Array3(v, s) => Ok((v.clone(), s.clone())),
@@ -242,11 +242,11 @@ impl Recorder for TensorboardRecorder {
                     RecordValue::DateTime(_) => {} // discard value
                     RecordValue::Array2(data, shape) => {
                         let shape = [3, shape[0], shape[1]];
-                        let min = data.iter().fold(f32::MAX, |m, v| v.min(m));
-                        let scale = data.iter().fold(-f32::MAX, |m, v| v.max(m)) - min;
+                        let min = data.iter().fold(f64::MAX, |m, v| v.min(m));
+                        let scale = data.iter().fold(-f64::MAX, |m, v| v.max(m)) - min;
                         let mut data = data
                             .iter()
-                            .map(|&e| ((e - min) / scale * 255f32) as u8)
+                            .map(|&e| ((e - min) / scale * 255f64) as u8)
                             .collect::<Vec<_>>();
                         let data_ = data.clone();
                         data.extend(data_.iter());
