+++ border-core/src/replay_buffer/base/iw_scheduler.rs
@@ -5,10 +5,10 @@ use serde::{Deserialize, Serialize};
 #[derive(Deserialize, Serialize, Clone, Debug, PartialEq)]
 pub struct IwScheduler {
     /// Initial value of $\beta$.
-    pub beta_0: f32,
+    pub beta_0: f64,
 
     /// Final value of $\beta$.
-    pub beta_final: f32,
+    pub beta_final: f64,
 
     /// Optimization steps when beta reaches its final value.
     pub n_opts_final: usize,
@@ -19,18 +19,18 @@ pub struct IwScheduler {
 
 impl IwScheduler {
     /// Creates a scheduler.
-    pub fn new(beta_0: f32, beta_final: f32, n_opts_final: usize) -> Self {
+    pub fn new(beta_0: f64, beta_final: f64, n_opts_final: usize) -> Self {
         Self { beta_0, beta_final, n_opts_final, n_opts: 0 }
     }
 
     /// Gets the exponents of importance sampling weight.
-    pub fn beta(&self) -> f32 {
+    pub fn beta(&self) -> f64 {
         let n_opts = self.n_opts;
         if n_opts >= self.n_opts_final {
             self.beta_final
         } else {
             let d = self.beta_final - self.beta_0;
-            self.beta_0 + d * (n_opts as f32 / self.n_opts_final as f32)
+            self.beta_0 + d * (n_opts as f64 / self.n_opts_final as f64)
         }
     }
