+++ border-core/src/replay_buffer/base/sum_tree.rs
@@ -21,33 +21,33 @@ pub enum WeightNormalizer {
 
 #[derive(Debug)]
 pub struct SumTree {
-    eps: f32,
-    alpha: f32,
+    eps: f64,
+    alpha: f64,
     capacity: usize,
     n_samples: usize,
-    tree: Vec<f32>,
-    min_tree: SegmentPoint<f32, MinIgnoreNaN>,
-    max_tree: SegmentPoint<f32, MaxIgnoreNaN>,
+    tree: Vec<f64>,
+    min_tree: SegmentPoint<f64, MinIgnoreNaN>,
+    max_tree: SegmentPoint<f64, MaxIgnoreNaN>,
     normalize: WeightNormalizer,
     _rng: fastrand::Rng,
 }
 
 impl SumTree {
-    pub fn new(capacity: usize, alpha: f32, normalize: WeightNormalizer) -> Self {
+    pub fn new(capacity: usize, alpha: f64, normalize: WeightNormalizer) -> Self {
         Self {
             eps: 1e-8,
             alpha,
             capacity,
             n_samples: 0,
-            tree: vec![0f32; 2 * capacity - 1],
-            min_tree: SegmentPoint::build(vec![f32::MAX; capacity], MinIgnoreNaN),
-            max_tree: SegmentPoint::build(vec![1e-8f32; capacity], MaxIgnoreNaN),
+            tree: vec![0f64; 2 * capacity - 1],
+            min_tree: SegmentPoint::build(vec![f64::MAX; capacity], MinIgnoreNaN),
+            max_tree: SegmentPoint::build(vec![1e-8f64; capacity], MaxIgnoreNaN),
             normalize,
             _rng: fastrand::Rng::with_seed(0),
         }
     }
 
-    fn propagate(&mut self, ix: usize, change: f32) {
+    fn propagate(&mut self, ix: usize, change: f64) {
         let parent = (ix - 1) / 2;
         self.tree[parent] += change;
         if parent != 0 {
@@ -55,7 +55,7 @@ impl SumTree {
         }
     }
 
-    fn retrieve(&self, ix: usize, s: f32) -> usize {
+    fn retrieve(&self, ix: usize, s: f64) -> usize {
         let left = 2 * ix + 1;
         let right = left + 1;
 
@@ -63,18 +63,18 @@ impl SumTree {
             return ix;
         }
 
-        if s <= self.tree[left] || self.tree[right] == 0f32 {
+        if s <= self.tree[left] || self.tree[right] == 0f64 {
             return self.retrieve(left, s);
         } else {
             return self.retrieve(right, s - self.tree[left]);
         }
     }
 
-    pub fn total(&self) -> f32 {
+    pub fn total(&self) -> f64 {
         return self.tree[0];
     }
 
-    pub fn max(&self) -> f32 {
+    pub fn max(&self) -> f64 {
         self.max_tree
             .query(0, self.max_tree.len())
             .powf(1.0 / self.alpha)
@@ -83,7 +83,7 @@ impl SumTree {
     /// Add priority value at `ix`-th element in the sum tree.
     ///
     /// The alpha-th power of the priority value is taken when addition.
-    pub fn add(&mut self, ix: usize, p: f32) {
+    pub fn add(&mut self, ix: usize, p: f64) {
         debug_assert!(ix <= self.n_samples);
 
         self.update(ix, p);
@@ -94,7 +94,7 @@ impl SumTree {
     }
 
     /// Update priority value at `ix`-th element in the sum tree.
-    pub fn update(&mut self, ix: usize, p: f32) {
+    pub fn update(&mut self, ix: usize, p: f64) {
         debug_assert!(ix < self.capacity);
 
         let p = (p + self.eps).powf(self.alpha);
@@ -111,7 +111,7 @@ impl SumTree {
     }
 
     /// Get the maximal index of the sum tree where the sum of priority values is less than `s`.
-    pub fn get(&self, s: f32) -> usize {
+    pub fn get(&self, s: f64) -> usize {
         let ix = self.retrieve(0, s);
         debug_assert!(ix >= (self.capacity - 1));
         ix + 1 - self.capacity
@@ -121,17 +121,17 @@ impl SumTree {
     ///
     /// The weight is $w_i=\left(N^{-1}P(i)^{-1}\right)^{\beta}$
     /// and it will be normalized by $max_i w_i$.
-    pub fn sample(&mut self, batch_size: usize, beta: f32) -> (Vec<i64>, Vec<f32>) {
+    pub fn sample(&mut self, batch_size: usize, beta: f64) -> (Vec<i64>, Vec<f64>) {
         let p_sum = &self.total();
         let ps = (0..batch_size)
-            .map(|_| p_sum * fastrand::f32())
+            .map(|_| p_sum * fastrand::f64())
             .collect::<Vec<_>>();
         let indices = ps.iter().map(|&p| self.get(p)).collect::<Vec<_>>();
         let (ws, w_max_inv) = self.weights(&indices, beta);
 
         // debug
         // if self.n_samples % 100 == 0 || p_sum.is_nan() || w_max.is_nan() {
-        if p_sum.is_nan() || w_max_inv.is_nan() || ws.iter().sum::<f32>().is_nan() {
+        if p_sum.is_nan() || w_max_inv.is_nan() || ws.iter().sum::<f64>().is_nan() {
             println!("self.n_samples: {:?}", self.n_samples);
             println!("p_sum: {:?}", p_sum);
             println!("w_max_inv: {:?}", w_max_inv);
@@ -162,8 +162,8 @@ impl SumTree {
         println!("total = {}", self.total());
     }
 
-    fn weights(&self, ixs: &Vec<usize>, beta: f32) -> (Vec<f32>, f32) {
-        let n = self.n_samples as f32 / self.total();
+    fn weights(&self, ixs: &Vec<usize>, beta: f64) -> (Vec<f64>, f64) {
+        let n = self.n_samples as f64 / self.total();
         let ws = ixs
             .iter()
             .map(|ix| self.tree[ix + self.capacity - 1])
@@ -173,9 +173,9 @@ impl SumTree {
         // normalizer within all samples
         let w_max_inv = match self.normalize {
             WeightNormalizer::All => (n * self.min_tree.query(0, self.n_samples)).powf(beta),
-            WeightNormalizer::Batch => 1f32 / ws.iter().fold(0.0 / 0.0, |m, v| v.max(m)),
+            WeightNormalizer::Batch => 1f64 / ws.iter().fold(0.0 / 0.0, |m, v| v.max(m)),
         };
-        let ws = ws.iter().map(|w| w * w_max_inv).collect::<Vec<f32>>();
+        let ws = ws.iter().map(|w| w * w_max_inv).collect::<Vec<f64>>();
 
         (ws, w_max_inv)
     }
@@ -187,7 +187,7 @@ mod tests {
 
     #[test]
     fn test_sum_tree_odd() {
-        let data = vec![0.5f32, 0.2, 0.8, 0.3, 1.1, 2.5, 3.9];
+        let data = vec![0.5f64, 0.2, 0.8, 0.3, 1.1, 2.5, 3.9];
         let mut sum_tree = SumTree::new(8, 1.0, Batch);
         for ix in 0..data.len() {
             sum_tree.add(ix, data[ix]);
@@ -217,7 +217,7 @@ mod tests {
         // let (ixs, _) = sum_tree.sample(n_samples, 1.0);
         // debug_assert!(ixs.iter().all(|&ix| ix < data.len() as i64));
         // (0..5).for_each(|ix| {
-        //     let p = data[ix] / sum_tree.total() * (n_samples as f32);
+        //     let p = data[ix] / sum_tree.total() * (n_samples as f64);
         //     let n = ixs.iter().filter(|&&e| e == ix as i64).collect::<Vec<_>>().len();
         //     println!("ix={:?}: {:?} (p={:?})", ix, n, p);
         // })
